refactor(masonry): [tower] refactor tree into tower and workspace manager
